# Optimize ap_entropy, samp_entropy